[NUI][AT-SPI] Clean up code
Rename structures with event arguments.
Hide at-spi implementation.
