+++ src/Tizen.NUI.Components/Controls/AlertDialog.cs
@@ -580,6 +580,10 @@ namespace Tizen.NUI.Components
             base.Dispose(type);
         }
 
+        /// <summary>
+        /// AccessibilityGetName.
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
         protected override string AccessibilityGetName()
         {
             if (!String.IsNullOrEmpty(Title))
