+++ src/Tizen.NUI.Components/Controls/CheckBox.cs
@@ -43,7 +43,10 @@ namespace Tizen.NUI.Components
         /// <since_tizen> 8 </since_tizen>
         public CheckBox(ButtonStyle buttonStyle) : base(buttonStyle) { }
 
-
+        /// <summary>
+        /// Initialize AT-SPI object.
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public override void OnInitialize()
         {
             base.OnInitialize();
