+++ src/Tizen.NUI.Components/Controls/Popup.cs
@@ -816,6 +816,10 @@ namespace Tizen.NUI.Components
             btGroup = new ButtonGroup(this);
         }
 
+        /// <summary>
+        /// Initialize AT-SPI object.
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public override void OnInitialize()
         {
             base.OnInitialize();
@@ -823,6 +827,10 @@ namespace Tizen.NUI.Components
             AppendAccessibilityAttribute("sub-role", "Alert");
         }
 
+        /// <summary>
+        /// Informs AT-SPI bridge about the set of AT-SPI states associated with this object.
+        /// </summary>
+        [EditorBrowsable(EditorBrowsableState.Never)]
         protected override AccessibilityStates AccessibilityCalculateStates()
         {
             var states = base.AccessibilityCalculateStates();
